Renamed class from adapter to Favorite adapter
