+++ app/src/main/java/com/tz/sanga/moviestore/Adapters/FavoriteAdapter.java
@@ -3,6 +3,7 @@ package com.tz.sanga.moviestore.Adapters;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,29 +16,27 @@ import com.bumptech.glide.load.resource.drawable.GlideDrawable;
 import com.bumptech.glide.request.RequestListener;
 import com.bumptech.glide.request.target.Target;
 import com.tz.sanga.moviestore.Constants;
-import com.tz.sanga.moviestore.Model.MovieObjects;
+import com.tz.sanga.moviestore.Database.Local.FavoriteNote;
 import com.tz.sanga.moviestore.R;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-
-public class FavoriteAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>{
+public class FavoriteAdapter extends RecyclerView.Adapter <RecyclerView.ViewHolder>{
     private Context context;
-    private List<MovieObjects>movieObjects= Collections.emptyList();
+    private List<FavoriteNote> movieObjects= new ArrayList<>();
     private dataListener listener;
     private favoriteOnLongClickListener favoriteListener;
+    private static final String TAG = "TAG";
 
-    public FavoriteAdapter(Context context, dataListener listener, favoriteOnLongClickListener favoriteListener, List<MovieObjects> movieObjects) {
+    public FavoriteAdapter(Context context, dataListener listener, favoriteOnLongClickListener favoriteListener) {
         this.context = context;
         this.listener = listener;
         this.favoriteListener = favoriteListener;
-        this.movieObjects = movieObjects;
     }
-
 
     @Override
     @NonNull
@@ -50,11 +49,11 @@ public class FavoriteAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolde
     @Override
     public void onBindViewHolder(@NonNull RecyclerView.ViewHolder viewHolder, int i) {
         final MyHolder myHolder = (MyHolder)viewHolder;
-        myHolder.object = movieObjects.get(i);
-        myHolder.textView.setText(movieObjects.get(i).getTitle());
+        myHolder.favoriteNote = movieObjects.get(i);
+        myHolder.textView.setText(myHolder.favoriteNote.getTitle());
 
         Glide.with(context)
-                .load(Constants.getImageUrl() +movieObjects.get(i).getPath())
+                .load(Constants.getImageUrl() +myHolder.favoriteNote.getPath())
                 .listener(new RequestListener<String, GlideDrawable>() {
                     @Override
                     public boolean onException(Exception e, String model, Target<GlideDrawable> target, boolean isFirstResource) {
@@ -78,11 +77,21 @@ public class FavoriteAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolde
         return movieObjects.size();
     }
 
-    class MyHolder extends RecyclerView.ViewHolder implements View.OnClickListener, View.OnLongClickListener {
-        @BindView(R.id.view_text)TextView textView;
-        @BindView(R.id.img_view)ImageView imageView;
-        @BindView(R.id.loading_favorite_poster)ProgressBar progressBar;
-        private MovieObjects object;
+
+    public void setFavoriteList (List<FavoriteNote>movieObjects) {
+        this.movieObjects = movieObjects;
+        notifyDataSetChanged();
+    }
+
+    public FavoriteNote getFavoriteNoteAt(int position){
+        return movieObjects.get(position);
+    }
+
+    class MyHolder extends RecyclerView.ViewHolder implements View.OnClickListener, View.OnLongClickListener{
+        @BindView(R.id.view_text) TextView textView;
+        @BindView(R.id.img_view) ImageView imageView;
+        @BindView(R.id.loading_favorite_poster) ProgressBar progressBar;
+        private FavoriteNote favoriteNote;
 
         public MyHolder(@NonNull View itemView) {
             super(itemView);
@@ -93,22 +102,21 @@ public class FavoriteAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolde
 
         @Override
         public void onClick(View v) {
-            listener.onClickFavorite(object.getOverview(), object.getPath());
+            listener.onClickFavorite(favoriteNote.getPath());
+            Log.d(TAG, "onClick: " + favoriteNote.getId());
         }
-
         @Override
         public boolean onLongClick(View v) {
-            favoriteListener.onLongClickFavorite(object.getPath(), object.getTitle());
+            int position = getAdapterPosition();
+            favoriteListener.onLongClickFavorite(movieObjects.get(position), favoriteNote.getTitle());
             return false;
         }
     }
 
     public interface dataListener {
-        void onClickFavorite(String overView, String path);
+        void onClickFavorite(String path);
     }
-   public interface favoriteOnLongClickListener{
-        void onLongClickFavorite(String path, String title);
-   }
+    public interface favoriteOnLongClickListener{
+        void onLongClickFavorite(FavoriteNote favoriteNote, String title);
+    }
 }
-
-
